feat: Update transaction status to CLOSED
This commit adds a new enum value 'Closed' to the TransactionStatus enum in the 'enums' module. It also includes the necessary code changes to update the transaction status to 'CLOSED' in the 'update' method of the 'TransactionsController' and 'update' method of the 'TransactionsService'. This change allows for closing a transaction by modifying its status.