feat: Add transaction_notification_email column to transactions table
This commit adds a new column, transaction_notification_email, to the transactions table in the database. The column is of type VARCHAR(255) and allows for storing email addresses for transaction notifications. This change enables the association of email addresses with transactions, facilitating the sending of notifications to the appropriate recipients.