feat: Update transaction service to include account logo URL
The code changes in `transactions.service.ts` add the `transaction_logo_url` field to the `createTransactionDto` object in the `TransactionsService` class. This field is populated with the `account_logo_url` value from the associated account. This update allows for the inclusion of the account logo URL in transaction data, enhancing the user interface by displaying logos alongside transactions.